requirements.txt
blinker==1.9.0
click==8.2.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.3.0
openpyxl==3.1.5
pandas==2.3.0
pillow==11.2.1
PyMuPDF==1.26.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
requests==2.32.3
validators==0.28.3
